Fix strncpy off-by-one bug.
path_len is the length of path without the NUL terminator. Because strncpy
writes only n characters, it would not include the NUL terminator in the
path string in the write to file_path. This caused problems when SDL gave
us a pointer to non-zeroed memory in file_path.

Change the call to strncpy to write path_len + 1, which is both the size of
the file_path buffer and the size of path including the NUL terminator. Now
file_path will be correct.