Fix mutex problem.
The mutex was destroyed before the audio device was closed. Since the audio
callback uses the SYNC macro, it could still be doing SDL_LockMutex and
SDL_UnlockMutex calls on the invalid emu_mutex pointer. This results in
spurious crashes.

This patch moves the SDL_DestroyMutex call to after SDL_CloseAudioDevice so
this doesn't happen. It also guards the calls to SDL_*Mutex and other
functions with log messages to indicate when they fail.